+++ frontend/smart-parking/uni_modules/lime-echart/components/l-echart/utils.js
@@ -57,7 +57,7 @@ export function wrapTouch(event) {
 	}
 	return event;
 }
-export const devicePixelRatio = wx.getSystemInfoSync().pixelRatio
+export const devicePixelRatio = uni.getSystemInfoSync().pixelRatio
 // #endif
 // #ifdef APP-NVUE
 export function base64ToPath(base64) {